+++ web/frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@viamrobotics/remote-control",
-  "version": "0.1.31",
+  "version": "0.1.32",
   "license": "Apache-2.0",
   "files": [
     "src/components/**/*.vue",
