+++ web/frontend/src/components/camera/cameras-list.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type {
   commonApi,
-  Client
+  Client,
 } from '@viamrobotics/sdk';
 
 import Camera from './camera.vue';
@@ -40,66 +40,66 @@ const setupCamera = (cameraName: string) => {
 
 <template>
   <v-collapse
-  v-for="camera in resources"
-  :key="camera.name"
-  :title="camera.name"
-  class="camera"
-  data-parent="app"
+    v-for="camera in resources"
+    :key="camera.name"
+    :title="camera.name"
+    class="camera"
+    data-parent="app"
   >
-  <v-breadcrumbs
-  slot="title"
-  crumbs="camera"
-  />
-
-  <div class="flex flex-col gap-4 border-x border-b border-black p-4">
-  <v-switch
-    :label="camera.name"
-    :aria-label="openCameras[camera.name] ? `Hide Camera: ${camera.name}` : `View Camera: ${camera.name}`"
-    :value="openCameras[camera.name] ? 'on' : 'off'"
-    @input="setupCamera(camera.name)"
-  />
-
-  <div
-    v-if="openCameras[camera.name]"
-    class="flex flex-wrap items-end gap-2"
->
-    <v-select
-    v-model="refreshFrequency[camera.name]"
-    class="w-fit"
-    label="Refresh frequency"
-    aria-label="Refresh frequency"
-    :options="Object.keys(selectedMap).join(',')"
+    <v-breadcrumbs
+      slot="title"
+      crumbs="camera"
     />
 
-    <v-button
-    v-if="refreshFrequency[camera.name] !== 'Live'"
-    icon="refresh"
-    label="Refresh"
-    @click="triggerRefresh = !triggerRefresh"
-    />
-  </div>
+    <div class="flex flex-col gap-4 border-x border-b border-black p-4">
+      <v-switch
+        :label="camera.name"
+        :aria-label="openCameras[camera.name] ? `Hide Camera: ${camera.name}` : `View Camera: ${camera.name}`"
+        :value="openCameras[camera.name] ? 'on' : 'off'"
+        @input="setupCamera(camera.name)"
+      />
 
-  <Camera
-    v-show="openCameras[camera.name]"
-    :key="camera.name"
-    :camera-name="camera.name"
-    :parent-name="parentName"
-    :client="client"
-    :resources="resources"
-    :show-export-screenshot="true"
-    :refresh-rate="refreshFrequency[camera.name]"
-    :trigger-refresh="triggerRefresh"
-  />
+      <div
+        v-if="openCameras[camera.name]"
+        class="flex flex-wrap items-end gap-2"
+      >
+        <v-select
+          v-model="refreshFrequency[camera.name]"
+          class="w-fit"
+          label="Refresh frequency"
+          aria-label="Refresh frequency"
+          :options="Object.keys(selectedMap).join(',')"
+        />
 
-  <PCD
-    :key="camera.name"
-    :camera-name="camera.name"
-    :parent-name="parentName"
-    :client="client"
-    :resources="resources"
-    :show-switch="true"
-    :show-refresh="true"
-  />
-  </div>
+        <v-button
+          v-if="refreshFrequency[camera.name] !== 'Live'"
+          icon="refresh"
+          label="Refresh"
+          @click="triggerRefresh = !triggerRefresh"
+        />
+      </div>
+
+      <Camera
+        v-show="openCameras[camera.name]"
+        :key="camera.name"
+        :camera-name="camera.name"
+        :parent-name="parentName"
+        :client="client"
+        :resources="resources"
+        :show-export-screenshot="true"
+        :refresh-rate="refreshFrequency[camera.name]"
+        :trigger-refresh="triggerRefresh"
+      />
+
+      <PCD
+        :key="camera.name"
+        :camera-name="camera.name"
+        :parent-name="parentName"
+        :client="client"
+        :resources="resources"
+        :show-switch="true"
+        :show-refresh="true"
+      />
+    </div>
   </v-collapse>
 </template>
