+++ web/frontend/src/components/navigation.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 
+// @ts-nocheck TODO: fix typecheck errors in https://viam.atlassian.net/browse/RSDK-1897
 import { ref, onUnmounted } from 'vue';
 import { grpc } from '@improbable-eng/grpc-web';
 import { toast } from '../lib/toast';
