+++ web/frontend/src/lib/log.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck TODO: fix typecheck errors in https://viam.atlassian.net/browse/RSDK-1897
 export const rcLogConditionally = (req: unknown) => {
   if (window.rcDebug) {
     console.log('gRPC call:', req);
